requirements.txt
Package             Version
------------------- ---------
asgiref             3.8.1
certifi             2025.1.31
charset-normalizer  3.4.1
colorama            0.4.6
dj-database-url     2.3.0
Django              5.1.7
djangorestframework 3.15.2
gunicorn            23.0.0
idna                3.10
iniconfig           2.1.0
packaging           24.2
pip                 24.3.1
pluggy              1.5.0
psycopg2-binary     2.9.10
pytest              8.3.5
pytest-django       4.10.0
python-decouple     3.8
python-dotenv       1.0.1
requests            2.32.3
sqlparse            0.5.3
typing_extensions   4.12.2
tzdata              2025.2
urllib3             2.3.0
whitenoise          6.9.0
